refactor(chat): extract artifact image gallery
